requirements.txt
openai==1.35.3
numpy==1.26.4
python-dotenv==1.0.1